Substituted core module
