+++ src/app/modules/core/core.module.ts
@@ -1,0 +1,17 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { NgbModule } from '@ng-bootstrap/ng-bootstrap';
+import { ReactiveFormsModule } from '@angular/forms';
+import { RouterModule } from '@angular/router';
+import { NgBootstrapFormValidationModule } from 'ng-bootstrap-form-validation';
+
+import { HomeComponent } from './components/home/home.component';
+import { NavbarComponent } from './components/navbar/navbar.component';
+import { LoginComponent } from './components/login/login.component';
+
+@NgModule({
+  declarations: [HomeComponent, NavbarComponent, LoginComponent],
+  imports: [CommonModule, NgbModule, ReactiveFormsModule, RouterModule, NgBootstrapFormValidationModule],
+  exports: [HomeComponent, NavbarComponent, LoginComponent]
+})
+export class CoreModule {}
